+++ apps/manifest-demo/3008-webpack-host/@mf-types/js-entry-provider/apis.d.ts
@@ -1,3 +1,0 @@
-
-    export type RemoteKeys = 'js-entry-provider/Component';
-    type PackageType<T> = T extends 'js-entry-provider/Component' ? typeof import('js-entry-provider/Component') :any;