+++ apps/router-demo/router-host-vue3-2100/src/App.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>
+    <h3>Vue3 app</h3>
+    <router-link to="/">Home</router-link>
+    <span> | </span>
+    <router-link to="/remote1">Remote1</router-link>
+    <router-view></router-view>
+  </div>
+</template>
+
+<style scoped></style>