+++ packages/rspack/src/ModuleFederationPlugin.ts
@@ -10,6 +10,9 @@ import {
 import { StatsPlugin } from '@module-federation/manifest';
 import { ContainerManager, utils } from '@module-federation/managers';
 import { DtsPlugin } from '@module-federation/dts-plugin';
+import ReactBridgePlugin from '@module-federation/bridge-react-webpack-plugin';
+import path from 'node:path';
+import fs from 'node:fs';
 
 type ExcludeFalse<T> = T extends undefined | false ? never : T;
 type SplitChunks = Compiler['options']['optimization']['splitChunks'];
@@ -108,6 +111,20 @@ export class ModuleFederationPlugin implements RspackPluginInstance {
         // @ts-ignore
       }).apply(compiler);
     }
+
+    // react bridge plugin
+    const nodeModulesPath = path.resolve(compiler.context, 'node_modules');
+    const reactPath = path.join(
+      nodeModulesPath,
+      '@module-federation/bridge-react',
+    );
+
+    // Check whether react exists
+    if (fs.existsSync(reactPath)) {
+      new ReactBridgePlugin({
+        moduleFederationOptions: this._options,
+      }).apply(compiler);
+    }
   }
 
   private _patchChunkSplit(compiler: Compiler, name: string): void {
@@ -142,10 +159,15 @@ export class ModuleFederationPlugin implements RspackPluginInstance {
           }
 
           if (cacheGroup.chunks === 'all') {
-            cacheGroup.chunks = new RegExp(
-              `^(?!.*(${name}|${name}_partial)).*$`,
-              'g',
-            );
+            cacheGroup.chunks = (chunk) => {
+              if (
+                chunk.name &&
+                (chunk.name === name || chunk.name === name + '_partial')
+              ) {
+                return false;
+              }
+              return true;
+            };
             break;
           }
           if (cacheGroup.chunks === 'initial') {