+++ packages/sdk/src/dom.ts
@@ -27,12 +27,13 @@ export type CreateScriptHookReturn =
   | { script?: HTMLScriptElement; timeout?: number }
   | void;
 
-export function createScript(
-  url: string,
-  cb: (value: void | PromiseLike<void>) => void,
-  attrs?: Record<string, any>,
-  createScriptHook?: (url: string) => CreateScriptHookReturn,
-): { script: HTMLScriptElement; needAttach: boolean } {
+export function createScript(info: {
+  url: string;
+  cb?: (value: void | PromiseLike<void>) => void;
+  attrs?: Record<string, any>;
+  needDeleteScript?: boolean;
+  createScriptHook?: (url: string) => CreateScriptHookReturn;
+}): { script: HTMLScriptElement; needAttach: boolean } {
   // Retrieve the existing script element by its src attribute
   let script: HTMLScriptElement | null = null;
   let needAttach = true;
@@ -42,7 +43,7 @@ export function createScript(
   for (let i = 0; i < scripts.length; i++) {
     const s = scripts[i];
     const scriptSrc = s.getAttribute('src');
-    if (scriptSrc && isStaticResourcesEqual(scriptSrc, url)) {
+    if (scriptSrc && isStaticResourcesEqual(scriptSrc, info.url)) {
       script = s;
       needAttach = false;
       break;
@@ -52,9 +53,9 @@ export function createScript(
   if (!script) {
     script = document.createElement('script');
     script.type = 'text/javascript';
-    script.src = url;
-    if (createScriptHook) {
-      const createScriptRes = createScriptHook(url);
+    script.src = info.url;
+    if (info.createScriptHook) {
+      const createScriptRes = info.createScriptHook(info.url);
 
       if (createScriptRes instanceof HTMLScriptElement) {
         script = createScriptRes;
@@ -63,18 +64,19 @@ export function createScript(
         if (createScriptRes.timeout) timeout = createScriptRes.timeout;
       }
     }
-  }
-
-  if (attrs) {
-    Object.keys(attrs).forEach((name) => {
-      if (script) {
-        if (name === 'async' || name === 'defer') {
-          script[name] = attrs[name];
-        } else {
-          script.setAttribute(name, attrs[name]);
+    const attrs = info.attrs;
+    if (attrs) {
+      Object.keys(attrs).forEach((name) => {
+        if (script) {
+          if (name === 'async' || name === 'defer') {
+            script[name] = attrs[name];
+            // Attributes that do not exist are considered overridden
+          } else if (!script.getAttribute(name)) {
+            script.setAttribute(name, attrs[name]);
+          }
         }
-      }
-    });
+      });
+    }
   }
 
   const onScriptComplete = (
@@ -88,34 +90,41 @@ export function createScript(
       script.onerror = null;
       script.onload = null;
       safeWrapper(() => {
-        script?.parentNode && script.parentNode.removeChild(script);
+        const { needDeleteScript = true } = info;
+        if (needDeleteScript) {
+          script?.parentNode && script.parentNode.removeChild(script);
+        }
       });
       if (prev) {
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         const res = (prev as any)(event);
-        cb();
+        info?.cb?.();
         return res;
       }
     }
-    cb();
+    info?.cb?.();
   };
 
   script.onerror = onScriptComplete.bind(null, script.onerror);
   script.onload = onScriptComplete.bind(null, script.onload);
 
   timeoutId = setTimeout(() => {
-    onScriptComplete(null, new Error(`Remote script "${url}" time-outed.`));
+    onScriptComplete(
+      null,
+      new Error(`Remote script "${info.url}" time-outed.`),
+    );
   }, timeout);
 
   return { script, needAttach };
 }
 
-export function createLink(
-  url: string,
-  cb: (value: void | PromiseLike<void>) => void,
-  attrs: Record<string, string> = {},
-  createLinkHook?: (url: string) => HTMLLinkElement | void,
-) {
+export function createLink(info: {
+  url: string;
+  cb: (value: void | PromiseLike<void>) => void;
+  attrs: Record<string, string>;
+  needDeleteLink?: boolean;
+  createLinkHook?: (url: string) => HTMLLinkElement | void;
+}) {
   // <link rel="preload" href="script.js" as="script">
 
   // Retrieve the existing script element by its src attribute
@@ -128,8 +137,8 @@ export function createLink(
     const linkRef = l.getAttribute('ref');
     if (
       linkHref &&
-      isStaticResourcesEqual(linkHref, url) &&
-      linkRef === attrs['ref']
+      isStaticResourcesEqual(linkHref, info.url) &&
+      linkRef === info.attrs['ref']
     ) {
       link = l;
       needAttach = false;
@@ -139,23 +148,23 @@ export function createLink(
 
   if (!link) {
     link = document.createElement('link');
-    link.setAttribute('href', url);
-    link.setAttribute('crossorigin', 'anonymous');
+    link.setAttribute('href', info.url);
 
-    if (createLinkHook) {
-      const createLinkRes = createLinkHook(url);
+    if (info.createLinkHook) {
+      const createLinkRes = info.createLinkHook(info.url);
       if (createLinkRes instanceof HTMLLinkElement) {
         link = createLinkRes;
       }
     }
-  }
 
-  if (attrs) {
-    Object.keys(attrs).forEach((name) => {
-      if (link) {
-        link.setAttribute(name, attrs[name]);
-      }
-    });
+    const attrs = info.attrs;
+    if (attrs) {
+      Object.keys(attrs).forEach((name) => {
+        if (link && !link.getAttribute(name)) {
+          link.setAttribute(name, attrs[name]);
+        }
+      });
+    }
   }
 
   const onLinkComplete = (
@@ -168,16 +177,19 @@ export function createLink(
       link.onerror = null;
       link.onload = null;
       safeWrapper(() => {
-        link?.parentNode && link.parentNode.removeChild(link);
+        const { needDeleteLink = true } = info;
+        if (needDeleteLink) {
+          link?.parentNode && link.parentNode.removeChild(link);
+        }
       });
       if (prev) {
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         const res = (prev as any)(event);
-        cb();
+        info.cb();
         return res;
       }
     }
-    cb();
+    info.cb();
   };
 
   link.onerror = onLinkComplete.bind(null, link.onerror);
@@ -193,14 +205,19 @@ export function loadScript(
     createScriptHook?: (url: string) => CreateScriptHookReturn;
   },
 ) {
-  const { attrs, createScriptHook } = info;
+  const { attrs = {}, createScriptHook } = info;
   return new Promise<void>((resolve, _reject) => {
-    const { script, needAttach } = createScript(
+    const { script, needAttach } = createScript({
       url,
-      resolve,
-      attrs,
+      cb: resolve,
+      attrs: {
+        crossorigin: 'anonymous',
+        fetchpriority: 'high',
+        ...attrs,
+      },
       createScriptHook,
-    );
-    needAttach && document.getElementsByTagName('head')[0].appendChild(script);
+      needDeleteScript: true,
+    });
+    needAttach && document.head.appendChild(script);
   });
 }