+++ apps/node-dynamic-remote/CHANGELOG.md
@@ -4,6 +4,36 @@
 
 ### Patch Changes
 
+- @module-federation/node@2.2.12
+
+## null
+
+### Patch Changes
+
+- @module-federation/node@2.2.11
+
+## null
+
+### Patch Changes
+
+- @module-federation/node@2.2.10
+
+## null
+
+### Patch Changes
+
+- @module-federation/node@2.2.9
+
+## null
+
+### Patch Changes
+
+- @module-federation/node@2.2.8
+
+## null
+
+### Patch Changes
+
 - @module-federation/node@2.2.7
 
 ## null