+++ apps/runtime-demo/3008-runtime-remote/CHANGELOG.md
@@ -1,5 +1,52 @@
 # 3008-runtime-remote
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [88445e7]
+  - @module-federation/enhanced@0.2.1
+
+## 1.0.11
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.0
+
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [1e93c5e]
+  - @module-federation/enhanced@0.1.21
+
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [766de80]
+- Updated dependencies [d5ac542]
+  - @module-federation/enhanced@0.1.20
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [e1518e2]
+  - @module-federation/enhanced@0.1.19
+
+## 1.0.7
+
+### Patch Changes
+
+- @module-federation/enhanced@0.1.18
+
+## 1.0.6
+
+### Patch Changes
+
+- @module-federation/enhanced@0.1.17
+
 ## 1.0.5
 
 ### Patch Changes