+++ packages/manifest/CHANGELOG.md
@@ -1,5 +1,82 @@
 # @module-federation/manifest
 
+## 0.2.1
+
+### Patch Changes
+
+- 88445e7: Support getPublicPath in compiler plugins
+- e494f1a: fix: fix expose collect assets do not filter entry point cause load extra file
+- Updated dependencies [88445e7]
+  - @module-federation/dts-plugin@0.2.1
+  - @module-federation/sdk@0.2.1
+  - @module-federation/managers@0.2.1
+
+## 0.2.0
+
+### Patch Changes
+
+- Updated dependencies [804cc3b]
+- Updated dependencies [52bb94c]
+- Updated dependencies [f7a6d3f]
+  - @module-federation/dts-plugin@0.2.0
+  - @module-federation/sdk@0.2.0
+  - @module-federation/managers@0.2.0
+
+## 0.1.21
+
+### Patch Changes
+
+- Updated dependencies [a50c459]
+- Updated dependencies [88900ad]
+- Updated dependencies [4cddb8a]
+- Updated dependencies [6ee10aa]
+- Updated dependencies [1e93c5e]
+  - @module-federation/dts-plugin@0.1.21
+  - @module-federation/sdk@0.1.21
+  - @module-federation/managers@0.1.21
+
+## 0.1.20
+
+### Patch Changes
+
+- Updated dependencies [685c607]
+- Updated dependencies [34f9498]
+- Updated dependencies [e8e0969]
+- Updated dependencies [349c381]
+  - @module-federation/dts-plugin@0.1.20
+  - @module-federation/sdk@0.1.20
+  - @module-federation/managers@0.1.20
+
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [031454d]
+- Updated dependencies [b0a31a7]
+- Updated dependencies [a2bfb9b]
+  - @module-federation/sdk@0.1.19
+  - @module-federation/dts-plugin@0.1.19
+  - @module-federation/managers@0.1.19
+
+## 0.1.18
+
+### Patch Changes
+
+- 80af3f3: fix: get remoteEntry type from options
+- Updated dependencies [80af3f3]
+  - @module-federation/sdk@0.1.18
+  - @module-federation/dts-plugin@0.1.18
+  - @module-federation/managers@0.1.18
+
+## 0.1.17
+
+### Patch Changes
+
+- 4f22c3e: fix: add requiredVersion to manifest.shared
+  - @module-federation/sdk@0.1.17
+  - @module-federation/managers@0.1.17
+  - @module-federation/dts-plugin@0.1.17
+
 ## 0.1.16
 
 ### Patch Changes